improve entities

+++ src/StudyThink.Domain/Entities/Payments/Payment.cs
@@ -6,5 +6,5 @@ public class Payment : BaseEntity
     public string Type { get; set; }
     public string Status { get; set; }
     public string Description { get; set; }
-    public int CourseId { get; set; }
+    public long CourseId { get; set; }
 }
